requirements.txt
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-mock==3.14.0
